refactor: Change data type of the id of all entities into 'long'

+++ src/main/java/com/walletbyhei/model/Account.java
@@ -11,7 +11,7 @@ import lombok.ToString;
 @ToString
 @EqualsAndHashCode
 public class Account {
-  private int account_id;
+  private long account_id;
   private String first_name;
   private String last_name;
   private String password;
@@ -19,7 +19,7 @@ public class Account {
   private BigDecimal balance;
   private int currency_id;
 
-  public void setAccount_id(int account_id) {
+  public void setAccount_id(long account_id) {
     this.account_id = account_id;
   }
