+++ src/main/java/com/walletbyhei/model/Currency.java
@@ -10,6 +10,6 @@ import lombok.*;
 @EqualsAndHashCode
 
 public class Currency {
-  private int currency_id;
+  private long currency_id;
   private String currency_name;
 }
